Refactor Cart and Success
